// src/model/DatasetType.ts
/**
 * An enum holding each predefined dataset type like Rubik, Chess etc
 */
export enum DatasetType {
    Rubik = 'rubik',
    Chess = 'chess',
    Neural = 'neural',
    Go = 'go',
    Test = 'test',
    Story = 'story',
    Cohort_Analysis = 'cohort_analysis',
    Chem = 'chem',
    Trrack = 'trrack',
    None = 'none',
    Subsampled = 'subsampled',
    Experiments_Only = 'experiments_only',
    Subsampled_with_Sugestions = 'subsampled_with_suggestions',
    Experiments_Only_with_Sugestions = 'experiments_only_with_suggestions',
    Sound = 'sound'
}